The answer will be sent to the evaluation model to evaluate the answer, and save it with its score in the database

+++ database/migrations/2020_03_31_211321_create_answers_table.php
@@ -16,7 +16,7 @@ class CreateAnswersTable extends Migration
         Schema::create(config('database.tables.answers'), function (Blueprint $table) {
             $table->increments('id');
             $table->text('body');
-            $table->integer('score')->nullable();
+            $table->double('score')->nullable();
             $table->integer('question_id')->unsigned();
             $table->integer('interview_id')->unsigned()->nullable();
             $table->timestamps();
